requirements.txt
numpy
opencv-python
python-dotenv
streamlit
openai
langchain